patch 8.1.0395: compiler warning on 64-bit MS-Windows
Problem:    Compiler warning on 64-bit MS-Windows.
Solution:   Add type cast. (Mike Williams)

+++ src/diff.c
@@ -712,7 +712,7 @@ diff_write_buffer(buf_T *buf, diffin_T *din)
 
     // xdiff requires one big block of memory with all the text.
     for (lnum = 1; lnum <= buf->b_ml.ml_line_count; ++lnum)
-	len += STRLEN(ml_get_buf(buf, lnum, FALSE)) + 1;
+	len += (long)STRLEN(ml_get_buf(buf, lnum, FALSE)) + 1;
     ptr = lalloc(len, TRUE);
     if (ptr == NULL)
     {
